[change] - bookCar() method - done

+++ backend-service/src/main/java/com/sda/rentalloapp/domain/Car.java
@@ -51,7 +51,7 @@ public class Car {
     @Column(unique = true)
     String vin;
 
-    int pricePerDayInPolishGrosz;
+    long pricePerDayInPolishGrosz;
 
     boolean availability;
