+++ backend-service/src/main/java/com/sda/rentalloapp/dto/CarDto.java
@@ -28,7 +28,7 @@ public record CarDto(
 
         String vin,
 
-        int pricePerDayInPolishGrosz,
+        long pricePerDayInPolishGrosz,
 
         boolean availability,
